+++ FRONTEND/startup-nextjs-main/src/app/blog/page.tsx
@@ -14,10 +14,8 @@ const Blog = () => {
   return (
     <>
       <Breadcrumb
-        pageName="Nuestros Servicios"
-        description="En un esfuerzo por concientizar sobre la importancia de la fauna y flora, debemos comprender su papel vital en los ecosistemas. Los animales salvajes, como jaguares, yaguaretés y aves exóticas, merecen protección en su hábitat natural. La caza ilegal y la degradación del entorno amenazan su supervivencia. Al conocer más sobre ellos y promover prácticas sostenibles, podemos preservar la biodiversidad y garantizar un futuro saludable para todas las especies. 🌿🦁🌎
-
-        Además, un paseo en auto por estas áreas naturales puede ser una experiencia enriquecedora. Imagina recorrer caminos rodeados de vegetación exuberante, avistando aves coloridas y, con suerte, observando la majestuosidad de un jaguar o yaguareté en su hábitat. Sin embargo, es crucial que estos paseos se realicen de manera responsable, sin perturbar a los animales ni dañar el entorno. Así, podemos disfrutar de la belleza natural mientras contribuimos a su preservación."
+        pageName="Alojamiento"
+        description="poner posibles alojamientos."
       />
 
       <section className="pb-[120px] pt-[120px]">
@@ -33,65 +31,7 @@ const Blog = () => {
             ))}
           </div>
 
-          <div className="-mx-4 flex flex-wrap" data-wow-delay=".15s">
-            <div className="w-full px-4">
-              <ul className="flex items-center justify-center pt-8">
-                <li className="mx-1">
-                  <a
-                    href="#0"
-                    className="flex h-9 min-w-[36px] items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color transition hover:bg-primary hover:bg-opacity-100 hover:text-white"
-                  >
-                    Prev
-                  </a>
-                </li>
-                <li className="mx-1">
-                  <a
-                    href="#0"
-                    className="flex h-9 min-w-[36px] items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color transition hover:bg-primary hover:bg-opacity-100 hover:text-white"
-                  >
-                    1
-                  </a>
-                </li>
-                <li className="mx-1">
-                  <a
-                    href="#0"
-                    className="flex h-9 min-w-[36px] items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color transition hover:bg-primary hover:bg-opacity-100 hover:text-white"
-                  >
-                    2
-                  </a>
-                </li>
-                <li className="mx-1">
-                  <a
-                    href="#0"
-                    className="flex h-9 min-w-[36px] items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color transition hover:bg-primary hover:bg-opacity-100 hover:text-white"
-                  >
-                    3
-                  </a>
-                </li>
-                <li className="mx-1">
-                  <span className="flex h-9 min-w-[36px] cursor-not-allowed items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color">
-                    ...
-                  </span>
-                </li>
-                <li className="mx-1">
-                  <a
-                    href="#0"
-                    className="flex h-9 min-w-[36px] items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color transition hover:bg-primary hover:bg-opacity-100 hover:text-white"
-                  >
-                    12
-                  </a>
-                </li>
-                <li className="mx-1">
-                  <a
-                    href="#0"
-                    className="flex h-9 min-w-[36px] items-center justify-center rounded-md bg-body-color bg-opacity-[15%] px-4 text-sm text-body-color transition hover:bg-primary hover:bg-opacity-100 hover:text-white"
-                  >
-                    Next
-                  </a>
-                </li>
-              </ul>
-            </div>
-          </div>
+         
         </div>
       </section>
     </>
